Eric wip

+++ src/Controller/VotesController.php
@@ -17,8 +17,8 @@ class VotesController extends AppController {
      * @return void Redirects on successful add, renders view otherwise.
      */
     public function adminAdd() {
-        
-    $this->adminOnly();
+
+        $this->adminOnly();
 
         $vote = $this->Votes->newEntity();
 
@@ -26,23 +26,23 @@ class VotesController extends AppController {
             $vote = $this->Votes->patchEntity($vote, $this->request->data);
             if ($this->Votes->save($vote)) {
                 $this->Flash->success('The vote has been saved.');
-                return $this->redirect(['controller' => 'Admin' ,'action' => 'votes']);
+                return $this->redirect(['controller' => 'Admin', 'action' => 'votes']);
             } else {
                 $this->Flash->error('The vote could not be saved. Please, try again.');
             }
         }
-        
+
         $this->set('users', $this->Votes->Users->find('list', [
-                'idField' => 'id',
-                'valueField' => 'username'
-            ]));
+                    'idField' => 'id',
+                    'valueField' => 'username'
+        ]));
         $this->set('pictures', $this->Votes->Pictures->find('list'));
         $this->set('marks', $this->Votes->Marks->find('list'));
     }
-    
+
     public function adminEdit($id = null) {
 
-	$this->adminOnly();
+        $this->adminOnly();
 
         $vote = $this->Votes->get($id);
         if ($this->request->is(['patch', 'post', 'put'])) {
@@ -56,16 +56,16 @@ class VotesController extends AppController {
         }
         $this->set('vote', $vote);
         $this->set('users', $this->Votes->Users->find('list', [
-                'idField' => 'id',
-                'valueField' => 'username'
-            ]));
+                    'idField' => 'id',
+                    'valueField' => 'username'
+        ]));
         $this->set('pictures', $this->Votes->Pictures->find('list'));
         $this->set('marks', $this->Votes->Marks->find('list'));
     }
-    
+
     public function adminDelete($id = null) {
 
-	$this->adminOnly();
+        $this->adminOnly();
 
         $entity = $this->Votes->get($id);
         if ($this->Votes->delete($entity)) {
@@ -74,25 +74,33 @@ class VotesController extends AppController {
             $this->Flash->error('Failed!');
         }
 
-        return $this->redirect($this->referer());    
+        return $this->redirect($this->referer());
     }
-    
-    public function voteAdd () {
+
+    public function voteAdd() {
+
+        if (!$this->Users->isVerified($this->user_id)) {
+            $this->Flash->error('Please verify your email before voting.');
+            return $this->redirect($this->referer());
+        }
+
         // Verify user_id isn't spoofed.
         if ($this->request->data(['user_id']) != $this->user_id) {
             $this->Flash->error('Failed!');
-        return $this->redirect($this->referer());
+            return $this->redirect($this->referer());
         }
-        
+
         $oldVote = $this->Votes->checkMark($this->request->data(['user_id']), $this->request->data(['mark_id']));
-        
+
         if ($oldVote != null) {
             $entity = $this->Votes->get($oldVote);
             if ($this->Votes->delete($entity)) {
                 $this->Flash->success('Deleted old vote!');
-            } else {$this->Flash->error('Failed at deleting old vote!');}
+            } else {
+                $this->Flash->error('Failed at deleting old vote!');
+            }
         }
-        
+
         $vote = $this->Votes->newEntity();
 
         if ($this->request->is('post')) {
@@ -104,6 +112,6 @@ class VotesController extends AppController {
                 $this->Flash->error('The vote could not be saved. Please, try again.');
             }
         }
-        
     }
+
 }
